+++ NEWS.md
@@ -10,6 +10,9 @@
   is shorter than the maximal AR order `p`. If it is longer, the the last `p`
   values in it are used.  This has always been the case when `N` is not missing.
 
+- new function `mix_qf` computes conditional quantiles, analogous to `mix_pdf`
+  and `mix_cdf`.
+
 - numerous fixes and improvements.
 
 # mixAR 0.22.5 (CRAN)